refactor: Change userId parameter to long

+++ src/main/java/com/mycompany/SkySong/authentication/service/DeleteService.java
@@ -1,5 +1,5 @@
 package com.mycompany.SkySong.authentication.service;
 
 public interface DeleteService {
-    String deleteUser(int userId);
+    String deleteUser(long userId);
 }
